refactor: major refactor of entire project
BREAKING CHANGE:
