+++ .gitignore
@@ -151,3 +151,4 @@ Notes.txt
 *backtest-chart.html
 *candlestick.html
 .vscode
+run.py
