+++ autotrader/autobot.py
@@ -3,14 +3,12 @@ import importlib
 import traceback
 import pandas as pd
 from datetime import datetime, timezone
-from autotrader.autodata import AutoData
 from autotrader.strategy import Strategy
 from autotrader.utilities import DataStream
 from autotrader.brokers.trading import Order
 from autotrader.brokers.broker import Broker
 from typing import TYPE_CHECKING, Literal, Union
-from concurrent.futures import ThreadPoolExecutor
-from autotrader.brokers.broker_utils import BrokerUtils
+from autotrader.brokers.virtual.broker import Broker as VirtualBroker
 from autotrader.utilities import get_data_config, TradeAnalysis, get_logger
 
 if TYPE_CHECKING:
@@ -65,11 +63,6 @@ class AutoTraderBot:
         ------
         Exception
             When there is an error retrieving the instrument data.
-
-        Returns
-        -------
-        None
-            The trading bot will be instantiated and ready for trading.
         """
         # Type hint inherited attributes
         self._run_mode: Literal["continuous", "periodic"]
@@ -84,7 +77,6 @@ class AutoTraderBot:
         self._data_path_mapper: dict
         self._backtest_mode: bool
         self._data_directory: str
-        self._broker_utils: BrokerUtils
         self._papertrading: bool
         self._max_workers: int
         self._scan_mode: bool
@@ -97,6 +89,7 @@ class AutoTraderBot:
         self._data_start: datetime
         self._data_end: datetime
         self._dynamic_data: bool
+        self._req_liveprice: bool
 
         # Inherit user options from autotrader
         for attribute, value in autotrader_instance.__dict__.items():
@@ -107,7 +100,7 @@ class AutoTraderBot:
 
         # Assign local attributes
         self.instrument = instrument
-        self._broker = broker
+        self._broker: Broker = broker
 
         # Define execution framework
         if self._execution_method is None:
@@ -219,70 +212,66 @@ class AutoTraderBot:
 
         portfolio = strategy_config["WATCHLIST"] if trade_portfolio else False
 
+        # ~~~~~~~~~~~~~~~~~~~~~~~~ BELOW TO BE DELETED ~~~~~~~~~~~~~~~~~~~~~~~~
         # Get data feed configuration
-        data_config = get_data_config(
-            feed=self._feed,
-            global_config=self._global_config_dict,
-            environment=self._environment,
-        )
-
-        # Construct AutoData instance
-        # TODO - shouldn't be creating a new instance if brokers already have their own
-        # The broker is the same as a data feed... Consider consolidating autodata into the
-        # brokers.
-        self._get_data = AutoData(
-            data_config=data_config,
-            allow_dancing_bears=self._allow_dancing_bears,
-            home_currency=self._base_currency,
-        )
+        # self._data_config = get_data_config(
+        #     feed=self._feed,
+        #     global_config=self._global_config_dict,
+        #     environment=self._environment,
+        # )
 
         # Create instance of data stream object
-        stream_attributes = {
-            "instrument": self.instrument,
-            "feed": self._feed,
-            "data_filepaths": self._data_filepaths,
-            "quote_data_file": self._quote_data_file,
-            "auxdata_files": self._auxdata_files,
-            "strategy_params": self._strategy_params,
-            "get_data": self._get_data,
-            "data_start": self._data_start,
-            "data_end": self._data_end,
-            "portfolio": portfolio,
-            "data_path_mapper": self._data_path_mapper,
-            "data_dir": self._data_directory,
-            "backtest_mode": self._backtest_mode,
-        }
-        self.datastream: DataStream = self._data_stream_object(**stream_attributes)
+        # Only when using local data...
+        # TODO - I think the way this will have to work is: brokers get instantiated
+        # from autotrader, the datastream gets constructed and passed to the brokers
+        # (or something similar), and then here, only the broker data methods get called
+        # to get the latest strategy data.
+        # But then should every exchange have their own datastream object? Why dont the data
+        # methods just get put in there?
+        # Each exchange has its own data stream... dont need to do anything here to
+        # instantiate the datastream.
+        # self._stream_attributes = {
+        #     "instrument": self.instrument,
+        #     "feed": self._feed,
+        #     "data_filepaths": self._data_filepaths,
+        #     "quote_data_file": self._quote_data_file,
+        #     "auxdata_files": self._auxdata_files,
+        #     "strategy_params": self._strategy_params,
+        #     # "get_data": self._get_data,
+        #     "data_start": self._data_start,
+        #     "data_end": self._data_end,
+        #     "portfolio": portfolio,
+        #     "data_path_mapper": self._data_path_mapper,
+        #     "data_dir": self._data_directory,
+        #     "backtest_mode": self._backtest_mode,
+        # }
+        # self.datastream: DataStream = self._data_stream_object(**stream_attributes)
+        # ~~~~~~~~~~~~~~~~~~~~~~~~ ABOVE TO BE DELETED ~~~~~~~~~~~~~~~~~~~~~~~~
 
-        # Add instantiated datastream object to AutoData instance
-        if self._feed == "local":
-            # TODO - tidy this, add `add_datastream` method to AutoData
-            self._get_data._datastream = self.datastream
+        # TODO - need to initialise the broker with a cache of data, if backtesting over
+        # full dataset, or otherwise just the window specified in the strategy.
+        # Initialise broker datasets
+        # The block below needs to be tidied with data config handling.
 
-            # Also add to brokers
-            if isinstance(self._broker, dict):
-                for broker in self._broker.values():
-                    # TODO - verify works with other exchanges, or only virtual
-                    broker.autodata._datastream = self.datastream
-            else:
-                self._broker.autodata._datastream = self.datastream
-
-        # Initial data call
-        self._refresh_data(deploy_dt)
+        for instrument, brokers in self._instrument_to_broker.items():
+            for broker in brokers:
+                # TODO - what if non virtual brokers are used?
+                broker._initialise_data(
+                    **{
+                        "instrument": instrument,
+                        "data_start": self._data_start,
+                        "data_end": self._data_end,
+                        "granularity": interval,
+                    }
+                )
 
         # Build strategy instantiation arguments
         strategy_inputs = {
             "parameters": params,
-            "data": self._strat_data,
+            # "data": self._strat_data,
             "instrument": self.instrument,
+            "broker": self._broker,
         }
-
-        if strategy_config["INCLUDE_BROKER"]:
-            strategy_inputs["broker"] = self._broker
-            strategy_inputs["broker_utils"] = self._broker_utils
-
-        if strategy_config["INCLUDE_STREAM"]:
-            strategy_inputs["data_stream"] = self.datastream
 
         # Instantiate Strategy
         my_strat: Strategy = strategy(**strategy_inputs)
@@ -300,6 +289,7 @@ class AutoTraderBot:
         self._strategy.stop_trading = autotrader_instance._remove_instance_file
 
         # Assign strategy attributes for tick-based strategy development
+        # TODO - improve type hints using strategy base class
         if self._backtest_mode:
             self._strategy._backtesting = True
             self.trade_results = None
@@ -307,7 +297,6 @@ class AutoTraderBot:
             self._strategy._tick_data = True
 
     def __repr__(self):
-        # TODO - alter str for portfolio bots
         if isinstance(self.instrument, list):
             return "Portfolio AutoTraderBot"
         else:
@@ -316,152 +305,85 @@ class AutoTraderBot:
     def __str__(self):
         return "AutoTraderBot instance"
 
-    def _update(self, i: int = None, timestamp: datetime = None) -> None:
+    def _update(self, timestamp: datetime) -> None:
         """Update strategy with the latest data and generate a trade signal.
 
         Parameters
         ----------
-        i : int, optional
-            The indexing parameter used when running in periodic update mode.
-            The default is None.
-
         timestamp : datetime, optional
-            The timestamp parameter used when running in continuous update
-            mode. The default is None.
-
-        Returns
-        -------
-        None
-            Trade signals generated will be submitted to the broker.
+            The current update time.
         """
+        if self._backtest_mode or self._papertrading:
+            # Update virtual broker
+            self._update_virtual_broker(dt=timestamp)
 
-        if self._run_mode == "continuous":
-            # Running in continuous update mode
-            strat_data, current_bars, quote_bars, sufficient_data = self._check_data(
-                timestamp, self._data_indexing
-            )
-            strat_object = strat_data
+        # Call strategy for orders
+        strategy_orders = self._strategy.generate_signal(timestamp)
 
-        else:
-            # Running in periodic update mode
-            current_bars = {self.instrument: self.data.iloc[i]}
-            quote_bars = {self.instrument: self.quote_data.iloc[i]}
-            sufficient_data = True
-            strat_object = i
+        # Check and qualify orders
+        orders = self._check_orders(strategy_orders)
+        self._qualify_orders(orders)
 
-        # Check for new data
-        new_data = self._check_last_bar(current_bars)
-
-        if sufficient_data and new_data:
-            # There is a sufficient amount of data, and it includes new data
-            if self._backtest_mode or self._papertrading:
-                # Update virtual broker with latest price bars
-                self._update_virtual_broker(current_bars)
-
-            # Get strategy orders
-            strategy_orders = self._strategy.generate_signal(strat_object)
-
-            # Check and qualify orders
-            orders = self._check_orders(strategy_orders)
-            self._qualify_orders(orders, current_bars, quote_bars)
-
-            if not self._scan_mode:
-                # Submit orders
-                if self._max_workers is not None:
-                    workers = min(self._max_workers, len(orders))
-                else:
-                    workers = None
-                with ThreadPoolExecutor(max_workers=workers) as executor:
-                    futures = []
-                    for order in orders:
-                        try:
-                            order_time = current_bars[order.instrument].name
-                        except:
-                            if self._feed == "none":
-                                order_time = datetime.now(timezone.utc)
-                            else:
-                                order_time = current_bars[order.data_name].name
-
-                        # Submit order to relevant exchange
-                        futures.append(
-                            executor.submit(
-                                self._execution_method,
-                                broker=self._brokers[order.exchange],
-                                order=order,
-                                order_time=order_time,
-                            )
-                        )
-
-                # Check for exceptions
-                for f in futures:
-                    try:
-                        f.result()
-                    except Exception as e:
-                        traceback_str = "".join(traceback.format_tb(e.__traceback__))
-                        exception_str = (
-                            f"AutoTrader exception when submitting order: {e}"
-                        )
-                        print_str = exception_str + "\nTraceback:\n" + traceback_str
-                        self.logger.error(print_str)
-
-            if self._papertrading:
-                # Update virtual broker again to trigger any orders
-                self._update_virtual_broker(current_bars)
-
-            try:
-                current_time = current_bars[list(current_bars.keys())[0]].name.strftime(
-                    "%b %d %Y %H:%M:%S"
-                )
-            except:
-                current_time = datetime.now().strftime("%b %d %Y %H:%M:%S")
-            if len(orders) > 0:
-                for order in orders:
-                    direction = "long" if order.direction > 0 else "short"
-                    order_string = (
-                        f"{current_time}: {order.instrument} "
-                        + f"{direction} {order.order_type} order of "
-                        + f"{order.size} units placed."
+        if not self._scan_mode:
+            # Submit orders
+            for order in orders:
+                # Submit order to relevant exchange
+                try:
+                    self._execution_method(
+                        broker=self._brokers[order.exchange],
+                        order=order,
+                        order_time=timestamp,
                     )
-                    self.logger.info(order_string)
-            else:
-                self.logger.debug(
-                    f"{current_time}: No signal detected ({self.instrument})."
+                except Exception as e:
+                    traceback_str = "".join(traceback.format_tb(e.__traceback__))
+                    exception_str = f"AutoTrader exception when submitting order: {e}"
+                    print_str = exception_str + "\nTraceback:\n" + traceback_str
+                    self.logger.error(print_str)
+
+        # If paper trading, update virtual broker again to trigger any orders
+        if self._papertrading:
+            self._update_virtual_broker(dt=timestamp)
+
+        # Log message
+        current_time = timestamp.strftime("%b %d %Y %H:%M:%S")
+        if len(orders) > 0:
+            for order in orders:
+                direction = "long" if order.direction > 0 else "short"
+                order_string = (
+                    f"{current_time}: {order.instrument} "
+                    + f"{direction} {order.order_type} order of "
+                    + f"{order.size} units placed."
                 )
-
-            # Check for orders placed and/or scan hits
-            if int(self._notify) > 0 and not (self._backtest_mode or self._scan_mode):
-                for order in orders:
-                    self._notifier.send_order(order)
-
-            # Check scan results
-            if self._scan_mode:
-                # Report AutoScan results
-                if int(self._verbosity) > 0 or int(self._notify) == 0:
-                    # Scan reporting with no notifications requested
-                    if len(orders) == 0:
-                        print("{}: No signal detected.".format(self.instrument))
-
-                    else:
-                        # Scan detected hits
-                        print("Scan hits:")
-                        for order in orders:
-                            print(order)
-
-                if int(self._notify) > 0:
-                    # Notifications requested
-                    for order in orders:
-                        self._notifier.send_message(f"Scan hit: {order}")
-
+                self.logger.info(order_string)
         else:
-            # Suppress error in backtest mode
-            self.logger.error(
-                "\nThe strategy has not been updated as there is either "
-                + "insufficient data, or no new data. If you believe "
-                + "this is an error, try setting allow_dancing_bears to "
-                + "True, or set allow_duplicate_bars to True in "
-                + "AutoTrader.configure().\n"
-                + f"Sufficient data: {sufficient_data}."
+            self.logger.debug(
+                f"{current_time}: No signal detected ({self.instrument})."
             )
+
+        # Check for orders placed and/or scan hits
+        if int(self._notify) > 0 and not (self._backtest_mode or self._scan_mode):
+            # TODO - what is this conditional?
+            for order in orders:
+                self._notifier.send_order(order)
+
+        # Check scan results
+        if self._scan_mode:
+            # Report AutoScan results
+            if int(self._verbosity) > 0 or int(self._notify) == 0:
+                # Scan reporting with no notifications requested
+                if len(orders) == 0:
+                    print(f"{self.instrument}: No signal detected.")
+
+                else:
+                    # Scan detected hits
+                    print("Scan hits:")
+                    for order in orders:
+                        print(order)
+
+            if int(self._notify) > 0:
+                # Notifications requested
+                for order in orders:
+                    self._notifier.send_message(f"Scan hit: {order}")
 
     def _refresh_data(self, timestamp: datetime = None, **kwargs):
         """Refreshes the active Bot's data attributes for trading.
@@ -491,13 +413,20 @@ class AutoTraderBot:
         -------
         None:
             The up-to-date data will be assigned to the Bot instance.
-
         """
+        # TODO - this method is now obsolete - delete it
 
         timestamp = datetime.now(timezone.utc) if timestamp is None else timestamp
 
         # Fetch new data
         # TODO - rename - timestamp is a datetimte object
+
+        # How many brokers are there? There could be multiple
+        # What if I just didn't get candles here? Instead just pass timestamp to strategy,
+        # and allow them to call candles if it wants...
+        # or any other broker methods.
+        # self._broker.data_broker.get_candles
+
         data, multi_data, quote_data, auxdata = self.datastream.refresh(
             timestamp=timestamp
         )
@@ -642,18 +571,15 @@ class AutoTraderBot:
             # Return empty list
             return []
 
-    def _qualify_orders(
-        self, orders: list[Order], current_bars: dict, quote_bars: dict
-    ) -> None:
-        """Passes price data to order to populate missing fields."""
-
+    def _qualify_orders(self, orders: list[Order]) -> None:
+        """Prepare orders for submission."""
         for order in orders:
             # Get relevant broker
             broker: Broker = self._brokers[order.exchange]
 
             # Fetch precision for instrument
             try:
-                precision = broker._utils.get_precision(order.instrument)
+                precision = broker.get_precision(order.instrument)
             except Exception as e:
                 # Print exception
                 self.logger.error("AutoTrader exception when qualifying order:", e)
@@ -661,40 +587,21 @@ class AutoTraderBot:
                 # Skip this order
                 continue
 
+            # Determine current price to assign to order
             if self._feed != "none":
-                # Get order price from current bars
-                if self._req_liveprice:
-                    # Fetch current price
-                    liveprice_func = getattr(
-                        self._get_data, f"_{self._feed.lower()}_liveprice"
-                    )
-                    last_price = liveprice_func(order)
-                else:
-                    # Fetch pseudo-current price
-                    try:
-                        # Use instrument
-                        last_price = self._get_data._pseduo_liveprice(
-                            last=current_bars[order.instrument].Close,
-                            quote_price=quote_bars[order.instrument].Close,
-                        )
-                    except:
-                        # Use data name
-                        last_price = self._get_data._pseduo_liveprice(
-                            last=current_bars[order.data_name].Close,
-                            quote_price=quote_bars[order.data_name].Close,
-                        )
+                # Get order price from current orderbook
+                orderbook = broker.get_orderbook(
+                    instrument=order.instrument,
+                )
 
-                if order.order_type not in ["close", "reduce", "modify"]:
-                    if order.direction < 0:
-                        order_price = last_price["bid"]
-                        HCF = last_price["negativeHCF"]
-                    else:
-                        order_price = last_price["ask"]
-                        HCF = last_price["positiveHCF"]
+                # Check order type to assign variables
+                # TODO - review use of HCF
+                if order.direction < 0:
+                    order_price = orderbook.bids.loc[0]["price"]
+                    HCF = 1
                 else:
-                    # Close, reduce or modify order type, provide dummy inputs
-                    order_price = last_price["ask"]
-                    HCF = last_price["positiveHCF"]
+                    order_price = orderbook.asks.loc[0]["price"]
+                    HCF = 1
 
             else:
                 # Do not provide order price yet
@@ -704,22 +611,12 @@ class AutoTraderBot:
             # Call order to update
             order(broker=broker, order_price=order_price, HCF=HCF, precision=precision)
 
-    def _update_virtual_broker(self, current_bars: dict) -> None:
-        """Updates virtual broker with latest price data."""
-        # TODO - the conditional here should allow specifically updating by L1,
-        # not only when feed=none
-        if self._feed == "none":
-            # None data feed provided, use L1 to update
-            for instrument, brokers in self._instrument_to_broker.items():
-                for broker in brokers:
-                    broker._update_instrument(instrument)
-
-        else:
-            # Using OHLC data feed
-            for product, bar in current_bars.items():
-                brokers = self._instrument_to_broker[product]
-                for broker in brokers:
-                    broker._update_positions(instrument=product, candle=bar)
+    def _update_virtual_broker(self, dt: datetime) -> None:
+        """Updates the virtual broker state. Only called when backtesting or paper trading."""
+        for instrument, brokers in self._instrument_to_broker.items():
+            for broker in brokers:
+                broker: VirtualBroker
+                broker._update_positions(instrument=instrument, dt=dt)
 
     def _create_trade_results(self, broker_histories: dict) -> dict:
         """Constructs bot-specific trade summary for post-processing."""
@@ -727,7 +624,7 @@ class AutoTraderBot:
         trade_results.indicators = (
             self._strategy.indicators if hasattr(self._strategy, "indicators") else None
         )
-        trade_results.data = self.data
+        trade_results.data = self._broker._data_cache[self.instrument]
         trade_results.interval = self._strategy_params["granularity"]
         self.trade_results = trade_results
 
@@ -736,6 +633,7 @@ class AutoTraderBot:
         the entire dataset is iterated over. For livetrading, only the latest candle
         is used. ONLY USED IN BACKTESTING NOW.
         """
+        # TODO - this method is now obsolete - delete it
         start_range = self._strategy_params["period"]
         end_range = len(self.data)
 
@@ -791,6 +689,7 @@ class AutoTraderBot:
             The checked data.
 
         """
+        # TODO - this method is now obsolete - delete it
         if check_for_future_data:
             if indexing.lower() == "open":
                 past_data = ohlc_data[ohlc_data.index < timestamp]
@@ -840,6 +739,7 @@ class AutoTraderBot:
         dict
             The checked auxiliary data.
         """
+        # TODO - this method is now obsolete - delete it
         processed_auxdata = {}
         for key, item in auxdata.items():
             if isinstance(item, pd.DataFrame) or isinstance(item, pd.Series):
@@ -860,10 +760,10 @@ class AutoTraderBot:
         Parameters
         ----------
         timestamp : datetime
-            DESCRIPTION.
+            Current update time.
 
         indexing : str, optional
-            DESCRIPTION. The default is 'open'.
+            Data reference point. The default is 'open'.
 
         Returns
         -------
@@ -880,6 +780,7 @@ class AutoTraderBot:
             Boolean flag whether sufficient data is available.
 
         """
+        # TODO - this method is now obsolete - delete it
 
         def get_current_bars(
             data: pd.DataFrame,
@@ -890,6 +791,8 @@ class AutoTraderBot:
             quote bars, then the quote_data boolean will be True.
             """
             if len(data) > 0:
+                # TODO - call broker to get orderbook instead of bars
+
                 current_bars = self.datastream.get_trading_bars(
                     data=data,
                     quote_bars=quote_data,
@@ -1029,6 +932,7 @@ class AutoTraderBot:
 
     def _check_last_bar(self, current_bars: dict) -> bool:
         """Checks for new data to prevent duplicate signals."""
+        # TODO - this method is now obsolete - delete it
         if self._allow_duplicate_bars:
             new_data = True
         else:
@@ -1074,7 +978,7 @@ class AutoTraderBot:
         if "plot_data" in strat_params and use_strat_plot_data:
             plot_data = strat_params["plot_data"]
         else:
-            plot_data = self.data
+            plot_data = self._broker.get_candles(instrument=self.instrument)
 
         return plot_data
