+++ autotrader/strategy.py
@@ -1,5 +1,7 @@
+from datetime import datetime
 from abc import ABC, abstractmethod
-from typing import List, Union, TYPE_CHECKING
+from autotrader.brokers.broker import Broker
+from typing import List, Union, TYPE_CHECKING, Optional
 
 if TYPE_CHECKING:
     from autotrader import Order
@@ -7,11 +9,13 @@ if TYPE_CHECKING:
 
 class Strategy(ABC):
     @abstractmethod
-    def __init__(self, parameters: dict, instrument: str, *args, **kwargs) -> None:
+    def __init__(
+        self, parameters: dict, instrument: str, broker: Broker, *args, **kwargs
+    ) -> None:
         super().__init__()
 
     @abstractmethod
-    def generate_signal(self, *args, **kwargs) -> Union["Order", List["Order"]]:
+    def generate_signal(self, dt: datetime) -> Optional[Union["Order", List["Order"]]]:
         """Generate trading signals based on the data supplied."""
 
     def stop_trading():
